One class per file in slashingProtection
