+++ packages/lodestar-validator/src/slashingProtection/attestation/index.ts
@@ -1,3 +1,4 @@
-export * from "./service";
-export * from "./dbRepository";
+export * from "./attestationByTargetRepository";
+export * from "./attestationLowerBoundRepository";
 export * from "./errors";
+export * from "./service";
