+++ packages/lodestar-validator/src/slashingProtection/block/blockBySlotRepository.ts
