+++ packages/lodestar-validator/src/slashingProtection/block/index.ts
@@ -1,3 +1,3 @@
-export * from "./service";
-export * from "./dbRepository";
+export * from "./blockBySlotRepository";
 export * from "./errors";
+export * from "./service";
