+++ packages/lodestar-validator/src/slashingProtection/minMaxSurround/distanceStoreRepository.ts
