+++ packages/lodestar-validator/src/slashingProtection/minMaxSurround/index.ts
@@ -1,4 +1,4 @@
-export * from "./interface";
-export * from "./dbRepository";
-export * from "./minMaxSurround";
+export * from "./distanceStoreRepository";
 export * from "./errors";
+export * from "./interface";
+export * from "./minMaxSurround";
